safari-534.53-branch/LayoutTests/http/tests/canvas/philip/tests/security.pattern.canvas.fillStyle.html

<!DOCTYPE html>
<title>Canvas test: security.pattern.canvas.fillStyle</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="security.pattern.image.fillStyle.html" title="security.pattern.image.fillStyle">
<link rel="next" href="security.pattern.image.strokeStyle.html" title="security.pattern.image.strokeStyle">
<body class="show_output">
<p>
 <a href="security.pattern.image.fillStyle.html" accesskey="p" title="[p] security.pattern.image.fillStyle">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="security.pattern.image.strokeStyle.html" accesskey="n" title="[n] security.pattern.image.strokeStyle">&gt;</a>
<h1><a href="index.security.html">security</a>.<a href="index.security.pattern.html">pattern</a>.<a href="index.security.pattern.canvas.html">canvas</a>.fillStyle</h1>
<p class="desc">Setting fillStyle to a pattern of an unclean canvas makes the canvas origin-unclean</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.security.fillStyle.canvas">security.fillStyle.canvas</a>
<li><a href="https://bugzilla.mozilla.org/show_bug.cgi?id=354127">Bugzilla</a>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var canvas2 = document.createElement('canvas');
canvas2.width = 100;
canvas2.height = 50;
var ctx2 = canvas2.getContext('2d');
ctx2.drawImage(document.getElementById('yellow.png'), 0, 0);
var p = ctx.createPattern(canvas2, 'repeat');
ctx.fillStyle = p;
ctx.fillStyle = 'red';
try { var _thrown = false; canvas.toDataURL(); } catch (e) { _thrown = true; } finally { _assert(_thrown, "should throw exception: canvas.toDataURL()"); }
try { var _thrown = false; ctx.getImageData(0, 0, 1, 1); } catch (e) { _thrown = true; } finally { _assert(_thrown, "should throw exception: ctx.getImageData(0, 0, 1, 1)"); }


});
</script>
<img src="http://localhost:8000/canvas/philip/images/yellow.png" id="yellow.png" class="resource">
