LayoutTests/http/tests/canvas/philip/tests/security.pattern.canvas.fillStyle-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
< [index] >
security.pattern.canvas.fillStyle
Setting fillStyle to a pattern of an unclean canvas makes the canvas origin-unclean
References: security.fillStyle.canvas Bugzilla
Actual output:
Passed
